feat: migration script 분리
