+++ package.json
@@ -14,14 +14,13 @@
     "build": "nest build",
     "start:dev": "nest start --watch",
     "migration": "npm run migration:local",
-    "migration:local": "dotenv -oe .env.local -- tsx ./scripts/migration.ts",
-    "migration:test": "dotenv -oe .env.test -- tsx ./scripts/migration.ts",
-    "migration:dev": "dotenv -oe .env.dev -- tsx ./scripts/migration.ts",
+    "migration:local": "dotenv -oe .env.local -- tsx ./scripts/migration",
+    "migration:test": "dotenv -oe .env.test -- tsx ./scripts/migration",
+    "migration:dev": "dotenv -oe .env.dev -- tsx ./scripts/migration",
     "ssh-tunnel": "env-cmd -e dev --use-shell tsx ./scripts/ssh-tunnel.ts",
     "format": "prettier --write **/* --ext .ts",
     "lint": "DEBUG=eslint:cli-engine eslint \"src/**/*.ts\"",
     "lint:fix": "DEBUG=eslint:cli-engine eslint --fix \"src/**/*.ts\"",
-    "lint:tsc": "tsc --noEmit",
     "test": "dotenv -oe .env.test -- jest",
     "coverage": "dotenv -oe .env.test -- jest --coverage",
     "update": "npx npm-check-updates -ui",
