+++ scripts/migration/commands/index.ts
@@ -1,0 +1,6 @@
+export { UpCommand } from './up';
+export { DownCommand } from './down';
+export { ListCommand } from './list';
+export { NewCommand } from './new';
+export { SeedCommand } from './seed';
+export { HelpCommand } from './help';
